// swiyu-ecosystem-portal-ui/src/app/api/generated/model/app-config.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {AuthConfig} from './auth-config';
import {EnvironmentsConfig} from './environments-config';
import {EportalConfig} from './eportal-config';
import {FeatureToggles} from './feature-toggles';
import {FunctionalityConfig} from './functionality-config';

/**
 * Application configuration for frontend
 */
export interface AppConfig {
  banner: string;
  portalUrl: string;
  productLabel: string;
  trustApiUrl: string;
  identifierRegistryApiUrl: string;
  statusRegistryApiUrl: string;
  apiGatewayAuthUrl: string;
  identifierRegistryUrl: string;
  productLabelColor: string;
  eportalUrl: string;
  tokenRefreshEnabled: boolean;
  maxBusinessPartnerPerCustomer: number;
  authConfig: AuthConfig;
  eportalConfig: EportalConfig;
  featureToggles: FeatureToggles;
  environments: EnvironmentsConfig;
  functionalityConfig: FunctionalityConfig;
}