// swiyu-ecosystem-portal-ui/src/app/api/generated/model/eportal-config.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {PamsEnvironment} from './pams-environment';

/**
 * E-Portal Service Navigation Widget configuration for frontend
 */
export interface EportalConfig {
  pamsAppId: string;
  infoContactEmail: string;
  pamsEnvironment: PamsEnvironment;
}
export namespace EportalConfig {}