// swiyu-ecosystem-portal-ui/src/app/api/generated/model/pams-environment.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const PamsEnvironment = {
  Dev: 'DEV',
  Ref: 'REF',
  Test: 'TEST',
  Abn: 'ABN',
  Prod: 'PROD'
} as const;
export type PamsEnvironment = (typeof PamsEnvironment)[keyof typeof PamsEnvironment];